Separate client to own file
